requirements.txt
Django==2.1.2
djangorestframework==3.9
djangorestframework-jwt==1.11.0
django-cors-middleware==1.3.1
django-bleach==0.3.0
django-enumfields==0.10.1
drf-nested-routers==0.91
pytz==2018.6
GitPython==2.1.11
lxml==4.2.0
pandas==0.23.4
hgapi==1.7.4
django-multi-email-field==0.5.1
django-background-tasks==1.2.0
scipy==1.2.1

